requirements.txt
Django
djangorestframework
django-rest-auth
django-cors-headers
psycopg2-binary
joblib
gunicorn
dj-database-url